Stats Multi in progress...

+++ FlexOfficeWebServer/flexoffice-adminui/src/main/java/com/orange/flexoffice/adminui/ws/endPoint/data/StatEndpoint.java
@@ -48,7 +48,7 @@ public interface StatEndpoint {
 	@GET
 	@Path(ROOMS_PATH + OCCUPANCY_PATH)
 	@Produces(MediaType.APPLICATION_JSON)
-	MultiStatSet getOccupancyStats(@QueryParam(FROM_PARAM) Integer from, @QueryParam(TO_PARAM) Integer to, @QueryParam(VIEW_TYPE_PARAM) String viewtype);
+	MultiStatSet getOccupancyStats(@QueryParam(FROM_PARAM) Long from, @QueryParam(TO_PARAM) Long to, @QueryParam(VIEW_TYPE_PARAM) String viewtype);
 		
 	// used for tests
 	boolean executeInitTestFile();
